rename: ComponentRender -> ComponentRenderer
