+++ src/lib/ComponentRenderer.svelte
