+++ src/lib/Render.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type {SvelteComponent} from 'svelte';
-	import ComponentRender from './ComponentRender.svelte';
+	import ComponentRenderer from './ComponentRenderer.svelte';
 	import type {RenderConfig} from './createRender';
 	import {isReadable, Undefined} from './store';
 
@@ -18,5 +18,5 @@
 {:else if typeof config !== 'object'}
 	{config}
 {:else}
-	<ComponentRender {config} />
+	<ComponentRenderer {config} />
 {/if}
